Add retries for RPC response read errors
During load spikes, snakebite sometimes fails to receive the
whole response from the NameNode. It then fails with a transient
exception. This enables the client to retry these errors to
make it more resilient.

Note: This retry as well as the already existing retries for
failovers/socket errors could potentially lead to duplicated
values being output by functions relying on Client._find_items
as it yields values on the go and a retry would duplicate yields.